move code locations
