+++ src/app/core/admin/admin-entity-details/admin-entity-form/admin-entity-form.component.ts
@@ -12,9 +12,9 @@ import {
 } from "@angular/cdk/drag-drop";
 import {
   ColumnConfig,
+  FormFieldConfig,
   toFormFieldConfig,
-} from "../../../common-components/entity-subrecord/entity-subrecord/entity-subrecord-config";
-import { FormFieldConfig } from "../../../common-components/entity-form/entity-form/FormConfig";
+} from "../../../common-components/entity-form/entity-form/FormConfig";
 import { AdminEntityService } from "../../admin-entity.service";
 import { lastValueFrom } from "rxjs";
 import { NgForOf, NgIf } from "@angular/common";
