+++ src/app/core/common-components/entities-table/list-paginator/list-paginator.component.scss
@@ -1,4 +1,4 @@
-@use "variables/sizes";
+@use "../../../../../styles/variables/sizes";
 
 // The paginator is usually a little smaller than this
 $approx-width-paginator: 450px;
