+++ src/app/core/common-components/entities-table/table-sort/table-sort.ts
@@ -1,7 +1,7 @@
-import { getReadableValue } from "./value-accessor";
+import { getReadableValue } from "../value-accessor/value-accessor";
 import { Entity } from "../../../entity/model/entity";
 import { Ordering } from "../../../basic-datatypes/configurable-enum/configurable-enum-ordering";
-import { TableRow } from "../../entities-table/entities-table.component";
+import { TableRow } from "../entities-table.component";
 
 /**
  * Custom sort implementation for a MatTableDataSource<TableRow<T>>
