+++ src/app/core/common-components/entity-field-edit/entity-field-edit.component.ts
@@ -6,8 +6,10 @@ import {
   EntityForm,
   EntityFormService,
 } from "../entity-form/entity-form.service";
-import { ColumnConfig } from "../entity-subrecord/entity-subrecord/entity-subrecord-config";
-import { FormFieldConfig } from "../entity-form/entity-form/FormConfig";
+import {
+  ColumnConfig,
+  FormFieldConfig,
+} from "../entity-form/entity-form/FormConfig";
 import { NgIf } from "@angular/common";
 import { EntityFieldViewComponent } from "../entity-field-view/entity-field-view.component";
