+++ src/app/core/common-components/entity-field-label/entity-field-label.component.ts
@@ -1,8 +1,10 @@
 import { Component, Input, OnChanges, SimpleChanges } from "@angular/core";
 import { MatTooltipModule } from "@angular/material/tooltip";
 import { EntityConstructor } from "../../entity/model/entity";
-import { ColumnConfig } from "../entity-subrecord/entity-subrecord/entity-subrecord-config";
-import { FormFieldConfig } from "../entity-form/entity-form/FormConfig";
+import {
+  ColumnConfig,
+  FormFieldConfig,
+} from "../entity-form/entity-form/FormConfig";
 import { EntityFormService } from "../entity-form/entity-form.service";
 import { NgIf } from "@angular/common";
