+++ src/app/core/filter/filter-generator/filter-predicate.ts
@@ -1,5 +1,5 @@
 import { Entity } from "../../entity/model/entity";
-import { getReadableValue } from "../../common-components/entity-subrecord/entity-subrecord/value-accessor";
+import { getReadableValue } from "../../common-components/entities-table/value-accessor/value-accessor";
 
 export function entityFilterPredicate(data: Entity, filter: string): boolean {
   return [...Object.values(data)].some((value) =>
