+++ src/app/core/filter/filter.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from "@angular/core";
 import { EntitySchemaField } from "../entity/schema/entity-schema-field";
-import { DataFilter } from "../common-components/entity-subrecord/entity-subrecord/entity-subrecord-config";
 import { Entity } from "../entity/model/entity";
 import {
   allInterpreters,
@@ -12,6 +11,7 @@ import {
 import moment from "moment";
 import { ConfigurableEnumService } from "../basic-datatypes/configurable-enum/configurable-enum.service";
 import { MongoQuery } from "@casl/ability";
+import { DataFilter } from "./filters/filters";
 
 /**
  * Utility service to help handling and aligning filters with entities.
