+++ src/app/core/form-dialog/form-dialog.service.ts
@@ -6,12 +6,12 @@ import {
 } from "@angular/material/dialog";
 import { ComponentType } from "@angular/cdk/overlay";
 import { Entity } from "../entity/model/entity";
-import { RowDetailsComponent } from "../common-components/entity-subrecord/row-details/row-details.component";
-import { FormFieldConfig } from "../common-components/entity-form/entity-form/FormConfig";
+import { RowDetailsComponent } from "./row-details/row-details.component";
 import {
   ColumnConfig,
+  FormFieldConfig,
   toFormFieldConfig,
-} from "../common-components/entity-subrecord/entity-subrecord/entity-subrecord-config";
+} from "../common-components/entity-form/entity-form/FormConfig";
 import { EntitySchemaService } from "../entity/schema/entity-schema.service";
 
 @Injectable({ providedIn: "root" })
